refactor(policy): move policy-controller library to runtime
The policy-controller main() has grown to hundreds of lines of business logic.
It's preferable to extract this into a library so that the business logic is
seperated from the process/runtime setup.
